Update registration after completed renewal (#187)
* Update registration after completed renewal

https://eaflood.atlassian.net/browse/WC-303

The user has entered all their details, paid their fees and not triggered any conviction checks. The renewal has reached the "Renewal complete" stage of the forms. Now it's time to actually renew the existing registration.

The end result should be that the registration is updated to include all information added or confirmed during the renewals process, with the previous version of the registration stored as a nested object. The exception is finance_details, which should always include all past orders and payments as well as the latest ones.

* Copy a registration to a past_registration

Before we update the registraton upon renewal, we should save a copy of the previous version of the registration.

This is then embedded in the registration as a past_registration.

* Start implementing RenewalCompletionService

This service will handle all the steps to update a registration when renewal is completed. Also add some tests to set out what the end result should be.

We have some issues with copying data around the registration - for example, from the registration to a nested past_registration, or a transient_registration to a registration. This is currently being fudged but will need further work.

* Fix how we copy attributes to past registration

Turns out this wasn't working as the registration we were copying from (and to!) had not been persisted.

* Persist past_registration when creating it

Trying to persist the past_registration in the service was causing issues, so save it on creation instead. But first we check to make sure a past_registration for this version of the registration doesn't already exists (to avoid duplication if we hit an error later in the process and have to start over).

* Update expiry date on renewal

The RenewalCompletionService now extends expires_on by another 3 years (or whatever time period the env var specifies).

* Destroy the transient_registration after renewal

Its work here is done.

* Copy data from transient_registration to registration

Now that we've collected our updated data, we should actually apply it. We don't copy over any data which is specific to the transient_registration (like the workflow_state or any temporary attributes).

* Replace values of unanswered questions with nil

If the user skips a question (for example, smart answers), the transient registration won't set a value for that attribute. But if they had an earlier answer which is saved to the registration, we want to overwrite that with a nil. Otherwise we risk old, irrelevant answers getting mixed in with the new data.

* Merge finance_details on renewal

We want to keep a record of existing payments and orders at the top level, so instead of hiding them in the past_registration, we keep them on the registration and merge in any new payments and orders from the transient_registration. Then we update the total balance.

* Renew registrations when user reaches renewal completion page

When a user is redirected to the renewal_complete_form after a successful payment, start the RenewalCompletionService to actually process the renewal. If the user tries to load the page but shouldn't be on it, the renewal will not happen.

* Fix how we check for existing past_registrations

There was a bug that meant that if there was already a backup, we create a new past_registration anyway, but without any attributes. This fixes that.

* Update metaData upon renewal

Make sure we set the correct route.

Also, update last_updated for registrations and transient_registrations any time we update them. (Not just during a renewal.)