Check renewal for relevant convictions (#193)
* Check renewal for relevant convictions

https://eaflood.atlassian.net/browse/WC-241

While users have the option to declare convictions themselves, we should also check details against our own conviction records and flag for review as required. The app should check key people and businesses. Any flagged renewals will be reviewed in the back office (implementation for this is a separate story).

If a renewal is flagged, users should be sent to the "renewal received" page, rather than "renewal complete", and the renewal should not be carried out until a member of staff reviews it in the back office.

* Set up EntityMatchingService

Create a new service to send entity matching requests to the Waste Carriers Service API. So far this just formats the correct URLs to send the query. For now, we return the URLs so we have something to test - this is just temporary.

This commit also sets up some extra date formatting, env vars and test factories to prepare for the request.

* Query the waste carriers service API

Use the URLs we've built to send an entity matching request to the waste carriers service API.

* Test error handling for entity match requests

Add some more tests to cover various errors we might get when making a request.

Also remove the :not_found response, as this was copied from the address lookup but doesn't apply here.

* If request hits an error, return unknown match data

This is to match the behaviour of the frontend app. If we get an error, we should record that the entity match result is unknown, so it can be flagged and checked again later.

* Save entity match result to transient_registration

If the result is for the business, it's stored on the transient_registration itself. If it's for a person, it's stored under that person.

We save the result whether it's a positive or negative match. If we run into an error when sending our query to the Java service, we save an "unknown" result.

* Run EntityMatchingService when declaration is posted

If a user posts the declaration form and the params are valid, we instantiate and run the EntiyMatchingService.

* Registrations with matches can't finish renewal immediately

If a business or a key person gets a positive or unknown match, don't renew that registration immediately, regardless of payment method. Instead, send the user to the "renewal application received" page. A member of staff can then check the matched result and confirm or reject the renewal in the back office.