Resolve issues with config found in testing (#202)
https://eaflood.atlassian.net/browse/WC-281

Following changes made to bring the renewals service env vars and config inline with all other waste carrier registration apps some issues were found.

These changes resolve the issues found in this project. Specifically

* Remove default values for registration charges

Having tried not setting these env vars and expecting the defaults to kick in, we realised that we had misunderstood how a command would work.

If an env var does not exist the call `ENV["MISSING_ENV_VAR"].to_i` will return 0, which means our default option is never used.

We could put in some code that first tests for the existance, but we felt its cleaner and simpler to just make sure the env var is set!

* Add byebug history to gitignore

When you use byebug it generates a file called `.byebug_history`. This change adds it to the gitignore file.

* Update example and travis period defaults

Spotted that the example .env file and the travis config file were using the old period of 6 months to renew a registration. When the service goes live that will be changed to 3 months so took this opportunity to update the values.