Migrate renewals journey to an engine (#203)
* Migrate renewals journey to an engine

https://eaflood.atlassian.net/browse/WC-188

The Waste Exemptions (WEX), Flood Risk Activity Exemptions (FRAE), and Project Application and Funding (PAF) services are built such that they are made of 3 projects:

- A front end external web app responsible just for styling and layout
- A back end external web app responsible for styling and layout, plus any additional back office functionality
- A shared 'engine' which contains all the logic and content for the registration/application journey

Waste carriers should follow the same model. To reach this, we need to rework the renewals journey into an engine, which can then be included in the front and back offices.

The following squashed commit history has been edited slightly (mainly to remove "changed X to Y, then changed Y back to X" commits), but changes generally fall into 3 categories:

- creating an engine
- moving functionality into the engine
- moving host application behaviour into a dummy app in /spec or the waste-carriers-front-office and -back-office host apps

* Set up gemspec

Create a gemspec for the new waste_carriers_engine, move in the dependencies and add a version for the engine.

* Move /app files to engine

Move controllers, forms, helpers, models, services and views which should be part of the engine to the folder. We don't make any changes to them yet as we want Git to recognise that the files have moved, not been deleted and recreated - that way we can hopefully keep the history.

Will this break the build? 100%.

* Move routes to engine, set up engine class

Move all the relevant routes into the engine module. Also create an Engine class which isolates the namespace.

* Move required classes and modules to WasteCarriersEngine

Move controllers, forms, models, helpers, services and validators to the WasteCarriersEngine.

The ability class also should not be part of the engine as it relates to the users, so that is moved back to the correct folder.

* Set up dummy app for testing

The dummy acts as a 'host' application for the engine, allowing us to test its behaviour as it would be in another app.

* Move specs to WasteCarriersEngine

* Start configuring dummy app

Move some config to the dummy app we use to test the engine. Also tweak some of the test config, and move some gems back from the gemspec to the Gemfile.

Some of the tests pass when you run them individually now, honest.

* Move concerns to autoloading folder

Our model concerns need to be in app/models/concerns/waste_carriers_engine, not app/models/waste_carriers_engine/concerns. Because.

* Tweak factory config to work with engine

To get factories to work, we specify the namespaced class they should be creating objects from. We also need to require factory_bot_rails in the rails_helper, rather than the Gemfile.

* Specify engine when using validators

Include the engine when we use our custom validator classes in the models or forms. Otherwise the app can't find them.

* Require gems for engine files which use them

We use some gems for making requests to APIs and for validation. However, even if they are defined in the gemspec, we still need to require them in the files which use them for them to work.

* Fix missing namespacing for engine classes in specs

Some specs and factories sit outside the engine. This meant that some mentions of classes within the engine weren't working - we fix this by changing Foo to WasteCarriersEngine::Foo.

* Get some request specs working

We do this by including the URL helpers in RSpec's rails helper, fixing some mentions of engine classes, and changing how we get the name of the form class in the FormsController to include the engine.

* Fix some asset loading issues

We need to include JQuery and Turbolinks in the Gemfile as they're needed for the dummy app. Also need to precompile the copy card photo.

* Move User and Ability into WasteCarriersEngine

After some discussion, decided to move functionality from Devise and CanCanCan into the engine and modify with config in host apps as needed.

It also turns out there's an easier way to add Devise helpers to request specs, so we can ditch RequestSpecHelper entirely.

* Define current_ability in ApplicationController

This fixes lots of the CanCanCan errors as it now correctly instantiates the Ability.

* Require Mongoid in engine.rb

CanCanCan insists that Mongoid is required before it, otherwise the adaptors don't work and some of the functionality is broken.

* Add tests for page title

These seem to have fallen out of our code coverage, so let's add some more tests.

Also discovered and fixed a bug where the ApplicationHelper specs failed if the file was run on its own, but passed if the entire suite was run.

* Move shared view partials to engine

Anything that's not part of the GDS frontend template and is part of the renewals journey instead should be moved into the engine.

The remaining non-engine files can be moved to the spec dummy or the waste-carriers-front-office app.

* Engine uses host app layout and custom Devise views

When the engine is mounted in a host app, it should use that app's default layout.

We also move the Devise views back up a level, as otherwise the custom views aren't used in the host app.

* Move Airbrake config to host app

To enable different host apps to have different Airbrake config, we move it out of the engine and into the spec dummy. This means we also need to allow for it not being installed in the host and avoid errors when this is the case.

* Move initialisers to dummy app

Secure headers and session store config should be part of the host app.

* Specify class name when storing Mongoid objects

Because of the namespacing, Mongoid attempted to store objects in different collections than we'd previously used. This meant it was unable to access data which had been added or seeded from another system. Specifying `store_in` fixes this issue.

The namespacing also interfered with Mongoid's ability to load embedded objects, so we add `class_name` to fix this.

* Enable HighVoltage pages to exist on engine and host app

Some static pages belong to the engine, while others belong to the host app. This commit makes some changes so they can coexist peacefully.

One thing I wasn't able to get working is using nested folders to create directories in the path - eg errors/invalid instead of just invalid. This may be something we come back to and try to iterate on later, but doesn't harm functionality at all, so isn't a pressing issue.

* Remove unnecessary initialisers

These either aren't doing anything or are covered by the host applications. Let's clear them out.

* Update README to cover engine

Our README should reflect that the repo is now for an engine rather than a standalone app. This commit adds some instructions for mounting it and removes some outdated info about running a Rails app.

We also remove the Hakiri badge as it's not used for gems. The front-office and back-office will be screened instead.

* Remove seeding

This no longer works as part of an engine, so we're relocating the seeds and the script to the waste-carriers-front-office app.

* Fix validation messages

Apps mounting the engine got translation errors when trying to display error messages. Turns out to be another namespacing issue in the language files - error text and form text are organised slightly differently.