Add missing fields to model (#219)
* Add reg_uuid field to model

This is used in the waste-carriers-frontend as part of the URL. To keep the two systems in sync, we should allow this on the engine's registration model.

* Add originalRegistrationNumber

Registrations which started out in IR have an original registration number which we need to track. This field isn't used in the renewal process in the engine at all, but we need to include it in the model so we can load ex-IR registrations into it.

* Add copy_cards field to registrations

Registrations from waste-carriers-frontend have an additional `copy_cards` field, as well as storing copy card info as OrderItems. For now, we just allow this field on the model so the two systems are compatible. We don't import it to the TransientRegistration as it's no longer relevant to this 'edition' of the registration.  However we will likely need additional changes in the future to make sure the behaviour of the field is the same as it is in the old system.