Add more missing attributes to registrations (#223)
* Add originalDateExpiry attribute to registrations

This attribute includes the original date of expiry for former IR registrations.

* Add accessCode attribute to registrations

This is used in the waste-carriers-frontend and waste-carriers-service.

* Alphabetise registration attributes and embedded objects

The list of attributes continues to grow and is getting disorderly. So let's just alphabetise it so it's easier to find or add stuff.

* Rename companyNo to company_no

waste-carriers-frontend calls this attribute companyNo. However the waste-carriers-service uses a snake_case version, and that is what is actually saved to the database.

We aliased this anyway, so luckily it's a quick fix.