'Add another user' does not return user to person form (#227)
https://eaflood.atlassian.net/browse/WC-414

When a user submits a person form with 'Add another user', the PersonController should redirect them back to the same form after saving. This is based on I18n text, which was changed when we namespaced everything for the engine and so broke. This commit fixes that.

The bug wasn't caught because we didn't update the I18n call in the controller and the spec, so both threw matching errors and then everything worked fine. To avoid this happening again, I've made the test more brittle by specifying the text it should expect.