Move users out of engine (#228)
* Move users out of engine

When we first set up the engine, we included users and related functionality like Devise and CanCanCan. However, it's become clear that the front office and back office users are sufficiently different that we shouldn't be trying to tackle this in the engine and should let the host apps handle it instead.

This PR strips out users, Devise and CanCanCan so they can be reimplemented in the host applications.

* Move user functionality from engine to spec dummy app

Cut out all the code for users, Devise or CanCanCan and shift it to the dummy app we use for RSpec testing.

We also remove the specs for users and sessions, as the host applications should handle this.

* Update README to include gem prerequisites

Since the engine expects Devise and CanCanCan to be installed, we should make this clear in the installation instructions.