Set order.updated_by_user to current_user (#237)
* Set order.updated_by_user to current_user

https://eaflood.atlassian.net/browse/WC-123

Currently this just reuses the account_email for the registration. However, in AD, the current_user would be a different user than the one who owns the registration, so we should correctly record who has modified it.

The Order.new_order method needs to have the current_user, so we have to pass that in from the controllers through to FinanceDetails.new_finance_details. Then we take the email address from the current_user.

This also involved a lot of changes to test data.