Set payment.updated_by_user to current_user (#239)
This previously just used the account email for the registration, but if the renewal is done through an assisted digital route, that won't be accurate. So we should use the current_user's email instead.

We fixed this issue for orders in https://github.com/DEFRA/waste-carriers-renewals/pull/237, but it turns out to affect payments as well.