Addresses don't allow firstName and lastName attributes (#240)
* Addresses don't allow firstName and lastName attributes

https://eaflood.atlassian.net/browse/WC-433

In waste-carriers-frontend, the firstName and lastName attributes are stored against the registration - however we also store them against the postal address. The engine doesn't allow this, which means it runs into issues when attempting to use addresses from the old system (for example, when updating a newly renewed registration).

We should allow these attributes to unblock renewals, but also make sure that we copy the firstName and lastName fields onto the postal address before renewing, so that we don't break any systems that rely on them (like mail merges).

* Add first_name and last_name to address attributes

This should stop the engine from freaking out when it touches an existing registration with these fields.

* Add names to contact address before renewing

We could add the names to the contact address when they are entered, or when the address is created – however this means we have to keep them in sync if the user goes back and modifies one or the other.

Since we don't use the names on the address at all during the renewal process, the safest option is just to copy them over as part of the renewal completion process.