Update last_modified before saving, not validation (#242)
Came across this issue while working on ordering in-progress renewals for WC-374. We call validations on registrations and transient_registrations on some occasions other than saving, which means that the last_modified field was being updated too often. Those changes weren't always saved, but they did cause problems when it came to ordering transient_registrations. So we should only update this field when we are actually modifying the object in the DB.