Add helper methods to determine status of transient_registration (#245)
https://eaflood.atlassian.net/browse/WC-374

In the back office, users need to be able to tell where a renewal is in the process and what actions are required to complete it. We started building in some of these checks in the back office, but actually it makes more sense to attach them directly to the object we are checking the status of.