Create conviction_sign_offs (#250)
* Create conviction_sign_offs

https://eaflood.atlassian.net/browse/WC-376

When a user has declared convictions or a matching conviction is found, the transient_registration should get a new conviction_sign_off object.

This is then used to determine whether a manual conviction check is required or not.

* Use ConvictionDataService to set up conviction info in DB

Previously we called the EntityMatchingService straight from the DeclarationFormController.

However, if a user declares they have a conviction, or if a match is found, we also want to create a new conviction_sign_off. Adding this to the EntityMatchingService would dilute its purpose too much. So instead we create a new service to call the EntityMatchingService and then set up data in the DB as required.

* Use conviction_sign_offs to determine if check is pending

Previously we used a combination of conviction_search_results and the value of declared_convictions to determine whether a check was required and pending.

However, the conviction_sign_off object will only be created if a check is required, so now we can just use that.

This is more consistent with how the waste-carriers-frontend operates.

* Rename spec file to include _spec at end

This was preventing RSpec from including the file when we ran `bundle exec rspec`, thus the mystery drop in test coverage.