Don't import revoked_reason to transient_registrations (#251)
This attribute is used to store reasons for approving or rejecting a registration if it was flagged with possible conviction matches. Because these convictions would be three years old at least, they are no longer relevant to the renewal, so we should not import the old revoked_reason as it's not related.