Error when renewing reg which no longer has conviction matches (#260)
https://eaflood.atlassian.net/browse/WC-455

This bug occurred when a registration which once had conviction matches (and therefore a conviction_sign_off) was renewed and no longer had any matches. We would expect the old conviction_sign_off to be removed but instead this produced a Mongoid error: `Updating the path 'conviction_sign_offs.0._id' would create a conflict at 'conviction_sign_offs'`

To fix this error, we make sure that we get the lists of in-use attribute keys from the registration and transient_registration _before_ we update the registration. Then we remove those attributes entirely instead of trying to set the value to nil.