Send email when renewal is completed (#261)
* Send email when renewal is completed

https://eaflood.atlassian.net/browse/WC-353

When a renewal is completed and the registration is updated, an email should be sent containing confirmation that the renewal has occurred, as well as any other relevant information.

The content of the email was confirmed in WC-195.

* Create RenewalMailer and basic email

`RenewalMailer.send_renewal_complete_email` now sends a basic email to the contact_email containing a test message.

Or would, if we had the SMTP settings configured.

* Set up renewal email template, including attached logo

We can reuse most of the template we use for sending out new registration emails, although the wording requires some tweaks to make it relevant. We also aren't attaching certificates at this time so this text needed to be removed.

The template itself also needed a lot of tidying up to make the code more readable and bring it up to our standards.

The code for attaching the image comes from here: https://github.com/DEFRA/waste-carriers-frontend/blob/master/app/helpers/email_helper.rb

* Improve email subject

Make the subject clearer and more specific.

* Send confirmation email after renewing

Plug the mailer into the RenewalCompletionService so the confirmation email is sent out after the renewal is completed.

This should work for both instant renewals and renewals which are initially delayed by payment, conviction checks, etc. Once the renewal is completed, the email will be sent.

* Use correct 'from' address when sending emails

This is set in an environment variable but stubbed for testing purposes.