Send email when renewal application is received and pending (#266)
* Send email when renewal application is received and pending

https://eaflood.atlassian.net/browse/WC-354

This task covers ensuring an email is sent when a user gets to the end of the renewal process but it is held as pending - this could be because of a pending payment or a conviction check. An email should be sent to confirm we have their details and the journey is complete.

* Send renewal received email after transitioning to that state

When AASM has a successful transition into renewal_received_form, the Mailer should send out the renewal_received email.

Currently the content of this email is the same as the 'renewal completed' one - the next step will be fixing that content so it's accurate.

* Send different emails depending on payment status

If payment is pending, we need to send an email which includes instructions for payment. However, if the balance has been paid, we send a different email which just says the application has been received.

Content for these emails is still to be added/amended, but we can now determine which template to use depending on the balance.

* Add relevant text to emails

The emails should now state what actions are required, if any, before the renewal is completed.

The pending payment one can reuse the text from the bank transfer page - this means that we only have to change that text in one location. The content on that page has also been through review so should be a better quality than the existing content.

We can remove @address_lines as it's not actually used for either renewal received email.

* Display datetime in human-readable format

* Improve styling, fix tests

Add some styling to the emails so they're a bit more readable and functional (especially tables with payment details).

Also improve our test factories to fix some failing tests.

* Display date_activated as a date

We store this as a DateTime to be compatible with existing data. However, the time is irrelevant so we should not display it.