Better renewal end screens for back office users (#270)
https://eaflood.atlassian.net/browse/WC-478

Currently the renewal_complete and renewal_received forms show the same content to all users regardless of which app they are mounted in.

Most of the time, this is fine - however, it does cause some problems on these two pages at the end of the renewals journey, as back office users require different links and information.

This change also allows the host application to override the engine's I18n translations. In the case of the frontend dashboard link it means the front office and back office can set it differently, and the engine doesn't have to know anything about the host application for this to work (so no logic to check if the engine is currently mounted in the back office or not).

We do this by modifying the load_path to include the engine's I18n files before the ones from the host app. The ones which are loaded last take precedence.