Implement separate controller actions for WorldPay responses (#272)
* Implement separate controller actions for WorldPay responses

https://eaflood.atlassian.net/browse/WC-409

Currently we have 2 different actions to deal with WorldPay responses - success or failure. However, there are actually 4 different responses from WorldPay that could lead to 'failure' - the payment could be pending, cancelled or throw an error as well as failing.

We should have separate controller actions to deal with each of these 5 scenarios. This will allow us to respond more appropriately to the different circumstances of a payment not immediately going through.

* Set up new controller actions for cancel, error and pending

Currently these behave exactly like the failure action.

* Use new actions in redirect URLs for WorldPay

Include the new URLs for the new actions in the URLs we give to WorldPay for redirecting back to our app.

* Validate payment statuses separately

We used to group all "failed" payment statuses together when validating. Now that they are separate actions, let's split them out.

* Call validations separately per action in WorldpayService

The WorldpayService is the connecting class between the controller and the validator, so we need separate methods to check if each action is valid.

* Check validation separately in controller

Each controller action now has its own valid_worldpay_[thing]_response? method.

* Reduce duplication in WorldpayFormsController

* Reduce duplication in WorldpayService

All the methods to check validity of unsuccessful payments were very similar, so we can move most of the logic to a single method and just pass in the validation method to use.

* Reduce more duplication in controller actions

Currently all responses to unsuccessful payments are very similar. So, like we did in the Service, we can move the logic to a single method and pass in the differentiating validation method.

* Remove unnecessary methods from controller

The `valid_worldpay_[action]_response?` are unnecessarily repetitive, so we can just move that logic back into the `success` and `respond_to_unsuccessful_payment` methods instead.

* Reduce duplication when checking paymentStatus validity

All these methods basically do the same thing, so we can just have a single method to check the status against a hash of valid options.

* Reduce duplication in public validation methods

Each response goes through the same validation steps, so we use a single method here rather than duplicating the calls.

* Remove duplicate [action]_path methods

These are all generated in the same way, so can be reduced to a single method.

* Remove duplicate [action]_url methods

Now these start to look very similar, so we can reduce to a single method with the action as an argument.

* Reduce duplication when testing unsuccessful responses

Move tests into a shared example.

* Reduce duplication in WasteCarriersService specs

We check the same things when testing the `valid_[action]?` methods for unsuccessful actions, so let's just have a single set of reusable specs.

* Move unsuccessful payment I18n messages to correct place

These should no longer all be stored under the failure page.

* Refactor respond_to_unsuccessful_payment for clarity

Move some logic into a separate `unsuccessful_response_is_valid?` method.

* Fix shared examples for unsuccessful payment requests

These all used the paths for a pending payment. Fixed to generate the path based on the action given as an argument.

* Fix names of valid pending paymentStatuses

'PENDING' is part of the URL but isn't actually a valid status. Fix this by adding the ones which are actually valid. Source: http://support.worldpay.com/support/kb/gg/corporate-gateway-guide/content/hostedintegration/paymentpageintegration.htm