Let users proceed to renewal_received after pending WP payment (#274)
* Let users proceed to renewal_received after pending WP payment

https://eaflood.atlassian.net/browse/WC-409

Sometimes WorldPay can't approve a payment immediately, although it may be approved after a delay.

In this case, we should let the user continue with the process as they will likely be charged eventually. However, as we don't want the renewal to be completed until payment has been taken, we should send it to the 'renewal received' state until a back office user records the payment.

* Transition to renewal_received after WorldPay if balance remains

If there is a positive balance left after going through WorldPay, for example if the payment hasn't been processed yet, redirect to renewal_received instead of renewal_complete.

* Send valid 'pending' WP responses to next page

Since there will still be a pending balance, the next page should be renewal_received_form.

* Get relevant order when calling worldpay_service method

We were creating a variable called `order` multiple times but only using it when creating a new WorldpayService. So we can reduce duplication by moving it to the method which creates the new WorldpayService.

Also rename that method to be a bit cleaner.

* Add pending_worldpay_payment? method

This returns true if the transient_registration has an order with a matching world_pay_status.

* Validate world_pay_status values in one place

Thanks to the new `pending_worldpay_payment?` method, we were validating whether a value was allowed for the type of WorldPay response in multiple places, so better to validate this in single reusable method, rather than introduce duplication.

* Better testing for Order.valid_world_pay_status?

Add specs which directly test the Order method. For other specs on other classes which rely on the method, stub the response so problems in the Order class don't break specs for other class.

* Use pending_worldpay_payment? for routing

Rather than checking if the balance is positive, use the new pending_worldpay_payment? method to determine if a user should be sent to renewal_received or renewal_complete.

* Send correct email when payment is pending

If a WorldPay payment is pending, email the user to let them know that we are processing the payment and they will receive a confirmation email once it has been received and renewal is completed.

* Display correct text on renewal_received if WorldPay payment pending

Let the user know that we are processing the payment and will confirm when it has been received and renewal has been completed.