Send WorldPay emails to contact_email for AD renewals (#278)
* Send WorldPay emails to contact_email for AD renewals

https://eaflood.atlassian.net/browse/WC-489

Users sometimes go through the assisted digital route to renew because they cannot access their account. When a user goes through WorldPay for payment, we send the account email through to WorldPay, which it then uses to send confirmation of the payment. If the user is AD because they can't access the account, this makes the confirmation meaningless.

So if a renewal is done via assisted digital, we should be passing the contact email to WorldPay, and not the account email.

* Add spec for AD renewals in WorldpayXmlService

The XML should use a different email address when the renewal is AD, so we need a new fixture and a new spec.

* WorldpayXmlService now takes a current_user when initialising

Improve the specs to handle this.

* Get different contact emails for digital and AD

We do this by checking if the current user's email matches the renewal account email. If they do, then the user who owns the registration is doing the renewal and it must be digital. If they don't, then another user is doing the renewal on their behalf and it must be assisted digital.